requirements.txt
torch
transformers
pillow
pytest
timm